Split steering into model, training and inference layers
The steering operation (direction, calibration, alignment, ReLU-gated
suppression, layer injection) is a definition on the model and is needed
by both the benchmark and the coming steering-training phase, so it moves
to odyssey/models/steering.py. The lifted-token builder is shared the same
way and moves to odyssey/training/lifted_tokens.py. The injection hook now
accepts any tensor that broadcasts against a block output, so a training
phase can push only the positions attributed to the concept.
